full session 17

+++ month02/week06/session17/01_session_17.sql
@@ -2,4 +2,4 @@ CREATE TABLE IF NOT EXISTS booklist (name TEXT, author TEXT, date TEXT);
 
 CREATE TABLE IF NOT EXISTS criticizer (name TEXT, location TEXT, count INT);
 
-CREATE TABLE IF NOT EXISTS termometer (temperature INT, date TEXT, time TEXT);
+CREATE TABLE IF NOT EXISTS termometer (temperature REAL, date TEXT, time TEXT);
